golden_ratio/requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.15.1
geographiclib==2.0
geonamescache==2.0.0
geopy==2.4.1
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
pillow==10.3.0
promise==2.3
psycopg2-binary==2.9.9
pycountry==24.6.1
PyJWT==1.7.1
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
